backend/pyproject.toml
[project]
name = "smartmart-backend"
version = "1.0.0"
description = "SmartMart Backend API Service with AI Recognition"
requires-python = ">=3.11"
dependencies = [
    # ========== 核心框架 ==========
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # ========== 数据库 ==========
    "sqlalchemy>=2.0.25",
    
    # ========== 文件处理 ==========
    "python-multipart>=0.0.6",
    "pillow>=10.2.0",
    
    # ========== WebSocket ==========
    "websockets>=12.0",
    
    # ========== AI 识别模块 ==========
    "numpy>=1.26.0",
    "faiss-cpu>=1.7.4",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    
    # ========== 工具 ==========
    "httpx>=0.25.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

# GPU 支持（可选）
gpu = [
    "faiss-gpu>=1.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

